src/third_party/blink/web_tests/external/wpt/offscreen-canvas/text/2d.text.draw.baseline.bottom-expected.txt

This is a testharness.js-based test.
FAIL textBaseline bottom is the bottom of the em square (not the bounding box) assert_approx_equals: Red channel of the pixel at (5, 5) expected 0 +/- 2 but got 255
Harness: the test ran to completion.
